requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alabaster==0.7.16
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
base58check==1.0.2
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
boto3==1.40.8
botocore==1.40.8
cachelib==0.13.0
cancelchain==1.4.1
celery==5.5.2
certifi==2023.11.17
cffi
channels==4.2.2
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
coverage==7.2.1
crispy-bootstrap5==2025.6
cron-descriptor==1.4.5
cryptography>=43.0.0
defusedxml==0.7.1
dj-database-url==2.3.0
dj-user-login-history==1.0.5
Django==4.2.23
django-admin-black==1.0.14
django-allauth==65.7.0
django-auth-style==0.8.1
django-browser-reload==1.17.0
django-celery-beat==2.8.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-cors-headers==4.3.1
django-countries==7.6.1
django-crispy-forms==2.4
django-debug-toolbar==5.0.1
django-easy-audit==1.3.5
django-extensions==3.2.3
django-filebrowser==4.0.3
django-grappelli==3.0.10
django-hosts==7.0.0
django-jazzmin==2.6.0
django-js-asset==3.1.2
django-localflavor==3.1
django-phonenumber-field==7.1.0
django-picklefield==3.3
django-q2==1.8.0
django-sms==0.7.0
django-tenants==3.8.0
django-timezone-field==7.0
django-tinymce==4.1.0
djangorestframework==3.14.0
djaodjin-deployutils==0.13.1
djaodjin-rules==0.4.8
docutils==0.19
dotenv==0.9.9
ecdsa==0.19.1
Envelopes==0.4
factory_boy==3.3.3
Faker==3.0.1
Flask==3.1.1
Flask-Caching==2.3.1
Flask-SQLAlchemy==3.0.5
frozenlist==1.6.0
greenlet==3.2.4
gunicorn==21.2.0
humanfriendly==10.0
icalendar==6.3.1
idna==3.6
imagesize==1.4.1
importlib-metadata==4.13.0
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.3
Markdown==3.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
messagebird==2.2.0
millify==0.1.1
multidict==6.4.3
oauthlib==3.2.2
packaging
passlib==1.7.4
pg8000==1.31.4
phonenumbers==8.13.7
pillow==11.2.1
pipdeptree==2.24.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2==2.9.10
pycparser==2.21
pycryptodome==3.23.0
Pygments==2.19.2
PyJWT==2.8.0
pymerkle==4.0.0
pytest==8.4.1
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-crontab==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-stdnum==2.1
python3-openid==3.2.0
pytz==2023.3.post1
rave-python==1.4.0
raven==6.10.0
razorpay==0.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rest-framework-generic-relations==2.2.0
rich==14.1.0
roman-numerals-py==3.1.0
s3transfer==0.13.1
schema==0.7.7
scramp==1.4.6
segno==1.6.6
sendsms==0.2.0
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.8
Sphinx==5.3.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-httpdomain==1.8.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-websupport==1.2.4
SQLAlchemy==1.4.54
sqlparse==0.5.3
stripe==5.0.0
text-unidecode==1.3
twilio==9.5.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.4.0
yarl==1.20.0
zipp==3.23.0
